Rip out the cog code generation for polarized structs and improve cross section calculation.
The cross section calculation wasn't minimizing calls to the structure functions as much as it could have been. In the process of rewriting this, it became apparent that each case had to be handled as a special case, i.e., the cog code generation that was meant to reduce duplication was not going to be useful here. So, the code generation for the polarized structs is ripped out entirely, as it is no longer helping with duplication and adds a lot of complexity to understanding the code.