Fix errors with how tagged unions interact with exceptions.
As they were before, the tagged unions with non-trivial destructors could accidentally double-call a destructor in certain circumstances. Specifically, when one variant is being swapped out for another, there is a brief moment where the tagged union is in an invalid state (the destructor for the old variant has already been called, but the constructor for the new variant has not yet been called). If an exception occurs in this moment, it will be bad. To fix this, add a tag that records whether the union is in a valid state or not.