Large refactor of the TMD-related classes to make all TMD return arrays for all flavours at once, instead of allowing individual flavors to be requested.
The reason behind this change is to make it easier to efficiently integrate with other libraries that provide TMDs. Often, these libraries will provide the TMDs all at once, instead of only providing the requested flavor. Also, this is preparation for caching of TMDs, which will significantly improve performance of Gaussian TMD convolutions.